+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All changes to macos-collector will be documented in this file.  
 
+## [1.1.0] - 2025-11-17
+### Added
+- DSStore File Collection (.DS_Store)
+- Unified Logs Collection
+
+## Fiex
+- Minor fixes and improvements
+
 ## [1.0.0] - 2025-10-30
 ### Added
 - Aftermath Collection